Many microscopic changes

+++ 2D/Belyi_Sphere.cpp
@@ -15,7 +15,7 @@ Stream <Hypermap> triangulations (int n) {
 		std::vector<Hypermap> hs;
 		for (auto alph : permutations(a)) {
 			Permutation alpha (n); for (int i=0; i<n-2*np; ++i) alpha[i] = alph[i];
-			for (long i=0; i<np; ++i) { alpha[n-2*np+2*i] = n-2*np+2*i+1; alpha[n-2*np+2*i+1] = n-2*np+2*i; }
+			for (int i=0; i<np; ++i) { alpha[n-2*np+2*i] = n-2*np+2*i+1; alpha[n-2*np+2*i+1] = n-2*np+2*i; }
 			if (!connected(phi,alpha)) continue;
 			Permutation sigma = (alpha*phi).inverse();
 			Hypermap h(sigma,alpha,phi);
